Move Annotation and AnnotationResource tests to adapter.dbpediaspotlight package
